+++ packages/elements/share/examples/04-integration.tsx
@@ -101,7 +101,6 @@ export default () => (
     originTracingFactory={() => mockOriginTracing}
     productId="confluence"
     shareAri="ari"
-    shareContentType="page"
     shareLink={window.location.href}
     shareTitle="My Share"
     triggerButtonStyle="icon-with-text"