+++ .changeset/b8d79153/changes.md
@@ -1,1 +1,2 @@
-- Exposed shareContentType prop to customise Share Form title
+- Exposed shareFormTitle prop to customise Share Form title
+- Removed object type from Copy link button